requirements.txt
certifi==2019.3.9
chardet==3.0.4
decorator==4.0.11
Django==2.2
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jsonapi==2.7.0
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.7
imageio==2.1.2
inflection==0.3.1
InstagramAPI==1.0.2
Markdown==3.1
moonmask==1.9
moviepy==0.2.3.2
numpy==1.16.2
Pillow==5.4.1
Pygments==2.3.1
PyJWT==1.7.1
pytz==2019.1
requests==2.20.0
requests-toolbelt==0.7.0
six==1.12.0
sqlparse==0.3.0
tqdm==4.11.2
urllib3==1.24.2
